+++ CHANGELOG.md
@@ -10,6 +10,10 @@
     - Disabled for specific code blocks with `using (new SuppressBlockingDetection())`
     - Doesn't detect everything. See original [Caveats described by Ben Adams](https://github.com/benaadams/Ben.BlockingDetector?tab=readme-ov-file#caveats).
 
+### Fixes
+
+- Empty strings are used instead of underscores to replace invalid metric tag values ([#3176](https://github.com/getsentry/sentry-dotnet/pull/3176))
+
 ### Dependencies
 
 - Bump Java SDK from v7.3.0 to v7.4.0 ([#3164](https://github.com/getsentry/sentry-dotnet/pull/3164))